Rename User class (to make errors)
I think it's better to replace User with ConnectionData.
Instead of storing all User in a vector, I'll just associate a
Connection with ConnectionData using a map.
There will be less searching this way.
I renamed the User class to UserX so I can see areas I need to work on.